+++ acm/DeleteOnContest.php
@@ -12,23 +12,34 @@ if ($conn->connect_error) {
     die("连接失败: " . $conn->connect_error);
 }
 
-$OnCNo = 100000000;
+$IsAdmin = false;
 
-$sql1 = "SELECT * FROM OnlineContest";
+if (isset($_COOKIE["user"])){
+    $PNo0 = $_COOKIE["user"];
+    $sql1 = "SELECT PAdmin FROM person WHERE PNo='$PNo0'";
+    $result = $conn->query($sql1);
+    $row = mysqli_fetch_assoc($result);
+    $IsAdmin = $row['PAdmin'];
+}
 
-$result = $conn->query($sql1);
+if ($IsAdmin) {
 
-$OnCNo = $OnCNo + $result->num_rows - 1;
-$OnCNo = "OnC".substr($OnCNo,1);
+    $OnCNo = $_POST['OnCNo'];
 
-$sql1 = "DELETE FROM OnlineContest WHERE OnCNo='$OnCNo'";
+    $sql1 = "DELETE FROM OnlineContest WHERE OnCNo='$OnCNo'";
 
-$result = $conn->query($sql1);
+    $result = $conn->query($sql1);
 
-if($result==TRUE) {
-    echo TRUE;
+    if ($result == TRUE) {
+        echo TRUE;
+    } else {
+        echo("错误描述: " . mysqli_error($conn));
+    }
+
 } else {
-    echo("错误描述: " . mysqli_error($conn));
+
+    echo "权限不足！";
+
 }
 
 $conn->close();
